feat: 添加事件仓储和快照仓储支持，优化类型系统
主要变更：
- 新增 EventRepository trait 用于事件存储抽象
- 新增 SnapshotRepository trait 用于快照存储抽象
- 将 version 类型从 i64 改为 usize（更符合语义）
- 为 Aggregate trait 添加 Serialize/Deserialize 约束
- 为 DomainEvent trait 添加完整的类型约束
- 简化 EventEnvelope 构造函数（移除 event_id 参数）
- 添加 serde/serde_json 依赖支持序列化

BREAKING CHANGES:
- version 字段类型从 i64 变更为 usize
- EventEnvelope::new() 移除了 event_id 参数
- Aggregate 和 DomainEvent 需要实现额外的 trait 约束

+++ ddd-macros/src/lib.rs
@@ -8,7 +8,7 @@ use syn::{
 };
 
 /// 聚合根宏
-/// 追加 id: IdType, version: i64 两个字段（若缺失）
+/// 追加 id: IdType, version: usize 两个字段（若缺失）
 /// 支持键值形式：#[aggregate(id = IdType)]，若不指定则默认为 String。
 #[proc_macro_attribute]
 pub fn aggregate(attr: TokenStream, item: TokenStream) -> TokenStream {
@@ -64,7 +64,7 @@ pub fn aggregate(attr: TokenStream, item: TokenStream) -> TokenStream {
     if let Some(f) = existed_version {
         new_named.push(f);
     } else {
-        new_named.push(syn::parse_quote! { version: i64 });
+        new_named.push(syn::parse_quote! { version: usize });
     }
 
     // 其他字段按原来顺序追加，但跳过 id/version，避免重复
@@ -89,7 +89,7 @@ pub fn aggregate(attr: TokenStream, item: TokenStream) -> TokenStream {
 /// pub enum XxxEvent {
 ///     Variant { field_a: T, ... },
 /// }
-/// 的具名字段变体，并为每个变体追加 id: IdType, version: i64 两个字段（若缺失）。
+/// 的具名字段变体，并为每个变体追加 id: IdType, version: usize 两个字段（若缺失）。
 /// 支持键值形式：#[event(id = IdType)]，若不指定则默认为 String。
 #[proc_macro_attribute]
 pub fn event(attr: TokenStream, item: TokenStream) -> TokenStream {
@@ -118,7 +118,7 @@ pub fn event(attr: TokenStream, item: TokenStream) -> TokenStream {
                     new_named.push(syn::parse_quote! { id: #id_type });
                 }
                 if !has_field_named(&fields_named, "version") {
-                    new_named.push(syn::parse_quote! { version: i64 });
+                    new_named.push(syn::parse_quote! { version: usize });
                 }
 
                 // 保留原有字段顺序
